Adding legacy stats db table
The legacy stats table contains rows for static legacy views and
download counts for each pid. These could be stats from a previous IR
or from Google Analytics prior to installing the Islandora Usage Stats
module. The legacy stats numbers are summed with the usage stats
numbers for the usage view block for an object record.